Move ToggleSwitch to dedicated directory
